Improved rendering and projection computation
* A more Oculus-SDK-like projection matrix is being used. It improves
the perceived image and degree of immersion. We apply the tracked
head information directly on cameras Postion, FocalPoint and ViewUp
vectors. Hence, tracking data and mouse data can be merged.

* Off-screen rendering is extended by multi-sampling enabled render
targets. Unfortunately, this comes along with a lot of hand written
OpenGL API calls. This needs to be re-factored in the future.

* Applying the chromatic aberration correction to the post process
shader.

Change-Id: I5009b32f5752b91f3f176445a435a75084d971a8